tidy up all the output::Markup code
